Move TahoeClient to own module, move http into TahoeClient class
